requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Environment & Configuration
python-dotenv>=1.0.0

# Authentication & Security
passlib>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.0.0

# Utilities
python-dateutil>=2.8.0
